storymate/requirements.txt
flask
python-dotenv
openai
langchain
langchain-core
langchain-community
chromadb
tiktoken
pymysql
json